backend/init/requirement.txt
backports.pbkdf2
datetime
flask
mysql-connector-python
PyJWT==1.7.1
requests
python-telegram-bot
